luci-app-pbr: Ignore WireGuard Server if on ignored list
This is the accompanying PR for: https://github.com/stangri/pbr/pull/29

This updates the help text, the read.me should probably also needs updating.

As this PR changes previous behavior I am not sure if this is the optimal solution, so feel free to change/improve  or discard this pull request.

It only updates the template, the translations should also need updating not sure if that is done automatically?

Signed-off-by: Erik Conijn <egc112@msn.com>